move top panel to own files
