+++ kahuna/public/js/components/gr-selection-top-bar/gr-selection-top-bar.html
@@ -1,0 +1,31 @@
+<gr-top-bar
+    class="gr-top-bar--blue gr-top-bar--overlay"
+    fixed="true"
+    ng:if="ctrl.active">
+    <gr-top-bar-nav gr:hide-home-link>
+        <button
+            class="top-bar-item clickable clickable--blue side-padded text-large"
+            type="button"
+            ng:click="ctrl.clearSelection()">
+            <gr-icon-label gr-icon="clear">Clear selection</gr-icon-label>
+        </button>
+        <div class="top-bar-item side-padded text-large">
+            {{ctrl.selectionCount}} selected
+        </div>
+    </gr-top-bar-nav>
+    <gr-top-bar-actions gr:hide-user-actions>
+        <gr-downloader
+            class="top-bar-item clickable clickable--blue side-padded"
+            gr:images="ctrl.selectedImages">
+        </gr-downloader>
+        <gr-delete-image
+            class="top-bar-item clickable clickable--blue side-padded"
+            images="ctrl.selectedImages"
+            ng:if="ctrl.selectionIsDeletable">
+        </gr-delete-image>
+        <gr-archiver
+            class="top-bar-item clickable clickable--blue side-padded"
+            gr:images="ctrl.selectedImages">
+        </gr-archiver>
+    </gr-top-bar-actions>
+</gr-top-bar>
